src/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-bootstrap4==2.3.1
django-filter==2.4.0
mysqlclient==2.0.3
Pillow==8.1.2
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
